# 1.57 ± 0.04 times faster