--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2018"
 [dependencies]
 flate2 = "1.0"
 structopt = "0.2"
+smol_str = "0.1.11"