--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 [dependencies]
 structopt = "0.2"
 smol_str = "0.1.11"
-hashbrown = "0.4.0"
+hashbrown = { version = "0.4.0", features = ["rayon"] }
 zstd = "0.4.24"
+rayon = "1.0.3"