… Cargo.lock
… lib/dictionary.txt.gz
… lib/dictionary.txt.zst