I have no idea how this works but I saw your post on Reddit mentioning performance and wanted to look into it a bit. This is the low hanging fruit.

```plain
hyperfine 'bin/wordsolve_before -m 3 -M 6 tootal' 'bin/wordsolve_latest -m 3 -M 6 tootal'
Benchmark #1: bin/wordsolve_before -m 3 -M 6 tootal
  Time (mean ± σ):     166.5 ms ±   3.9 ms    [User: 159.9 ms, System: 3.9 ms]
  Range (min … max):   163.3 ms … 178.2 ms
 
  Warning: The first benchmarking run for this command was significantly slower than the rest (178.2 ms). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.
 
Benchmark #2: bin/wordsolve_latest -m 3 -M 6 tootal
  Time (mean ± σ):     105.9 ms ±   1.2 ms    [User: 99.2 ms, System: 4.5 ms]
  Range (min … max):   103.9 ms … 110.1 ms
 
Summary
  'bin/wordsolve_latest -m 3 -M 6 tootal' ran
    1.57 ± 0.04 times faster than 'bin/wordsolve_before -m 3 -M 6 tootal'
```